refactor(evm-tx-data): docs, better var names, and DRY improvements
