+++ proto/eth/evm/v1/tx.proto
@@ -51,8 +51,12 @@ message MsgEthereumTx {
 }
 
 // LegacyTx is the transaction data of regular Ethereum transactions.
-// NOTE: All non-protected transactions (i.e non EIP155 signed) will fail if the
-// AllowUnprotectedTxs parameter is disabled.
+//
+// Note that setting "evm.Params.AllowUnprotectedTxs" to false will cause all
+// non-EIP155 signed transactions to fail, as they'll lack replay protection.
+//
+// LegacyTx is a custom implementation of "LegacyTx" from
+// "github.com/ethereum/go-ethereum/core/types".
 message LegacyTx {
   option (gogoproto.goproto_getters) = false;
   option (cosmos_proto.implements_interface) = "TxData";
@@ -69,15 +73,28 @@ message LegacyTx {
   string value = 5 [(gogoproto.customtype) = "cosmossdk.io/math.Int", (gogoproto.customname) = "Amount"];
   // data is the data payload bytes of the transaction.
   bytes data = 6;
-  // v defines the signature value
+
+  // v defines the recovery id as the "v" signature value from the elliptic curve
+  // digital signatute algorithm (ECDSA). It indicates which of two possible
+  // solutions should be used to reconstruct the public key from the signature.
+  // In Ethereum, "v" takes the value 27 or 28 for transactions that are not
+  // relay-protected.
   bytes v = 7;
-  // r defines the signature value
+
+  // r defines the x-coordinate of a point on the elliptic curve in the elliptic curve
+  // digital signatute algorithm (ECDSA). It's crucial in ensuring uniqueness of
+  // the signature.
   bytes r = 8;
-  // s define the signature value
+
+  // s define the signature value derived from the private key, message hash, and
+  // the value of "r". It ensures that the signature is tied to both the message
+  // and the private key of the sender.
   bytes s = 9;
 }
 
 // AccessListTx is the data of EIP-2930 access list transactions.
+// It is a custom implementation of "AccessListTx" from
+// "github.com/ethereum/go-ethereum/core/types".
 message AccessListTx {
   option (gogoproto.goproto_getters) = false;
   option (cosmos_proto.implements_interface) = "TxData";
@@ -103,15 +120,28 @@ message AccessListTx {
   // accesses is an array of access tuples
   repeated AccessTuple accesses = 8
       [(gogoproto.castrepeated) = "AccessList", (gogoproto.jsontag) = "accessList", (gogoproto.nullable) = false];
-  // v defines the signature value
+
+  // v defines the recovery id and "v" signature value from the elliptic curve
+  // digital signatute algorithm (ECDSA). It indicates which of two possible
+  // solutions should be used to reconstruct the public key from the signature.
+  // In Ethereum, "v" takes the value 27 or 28 for transactions that are not
+  // relay-protected.
   bytes v = 9;
-  // r defines the signature value
+
+  // r defines the x-coordinate of a point on the elliptic curve in the elliptic curve
+  // digital signatute algorithm (ECDSA). It's crucial in ensuring uniqueness of
+  // the signature.
   bytes r = 10;
-  // s define the signature value
+
+  // s define the signature value derived from the private key, message hash, and
+  // the value of "r". It ensures that the signature is tied to both the message
+  // and the private key of the sender.
   bytes s = 11;
 }
 
-// DynamicFeeTx is the data of EIP-1559 dinamic fee transactions.
+// DynamicFeeTx is the data of EIP-1559 dynamic fee transactions. It is a custom
+// implementation of "DynamicFeeTx" from
+// "github.com/ethereum/go-ethereum/core/types".
 message DynamicFeeTx {
   option (gogoproto.goproto_getters) = false;
   option (cosmos_proto.implements_interface) = "TxData";
@@ -139,11 +169,21 @@ message DynamicFeeTx {
   // accesses is an array of access tuples
   repeated AccessTuple accesses = 9
       [(gogoproto.castrepeated) = "AccessList", (gogoproto.jsontag) = "accessList", (gogoproto.nullable) = false];
-  // v defines the signature value
+  // v defines the recovery id and "v" signature value from the elliptic curve
+  // digital signatute algorithm (ECDSA). It indicates which of two possible
+  // solutions should be used to reconstruct the public key from the signature.
+  // In Ethereum, "v" takes the value 27 or 28 for transactions that are not
+  // relay-protected.
   bytes v = 10;
-  // r defines the signature value
+
+  // r defines the x-coordinate of a point on the elliptic curve in the elliptic curve
+  // digital signatute algorithm (ECDSA). It's crucial in ensuring uniqueness of
+  // the signature.
   bytes r = 11;
-  // s define the signature value
+
+  // s define the signature value derived from the private key, message hash, and
+  // the value of "r". It ensures that the signature is tied to both the message
+  // and the private key of the sender.
   bytes s = 12;
 }
